requirements.txt
moviepy
Pillow
openai-whisper
